+++ utilities/blob_db/blob_log_format.h
@@ -109,8 +109,8 @@ struct BlobLogRecord {
   // header include fields up to blob CRC
   static constexpr size_t kHeaderSize = 32;
 
-  uint64_t key_size = 0;
-  uint64_t value_size = 0;
+  size_t key_size = 0;
+  size_t value_size = 0;
   uint64_t expiration = 0;
   uint32_t header_crc = 0;
   uint32_t blob_crc = 0;
